// src/types/indexer/indexerManual.ts
export interface IndexerCompositeOrderObject {
  id: string;
  subaccountId?: string;
  clientId?: string;
  clobPairId?: string;
  side: IndexerOrderSide;
  size: string;
  totalFilled: string;
  price: string;
  type: IndexerOrderType;
  reduceOnly?: boolean;
  orderFlags?: string;
  goodTilBlock?: string | null;
  goodTilBlockTime?: string | null;
  createdAtHeight?: string | null;
  clientMetadata?: string;
  triggerPrice?: string | null;
  timeInForce?: IndexerAPITimeInForce;
  status?: IndexerAPIOrderStatus;
  postOnly?: boolean;
  ticker: string;
  updatedAt?: IndexerIsoString | null;
  updatedAtHeight?: string | null;
  subaccountNumber: number;
  removalReason?: string;
  totalOptimisticFilled?: string;
}
// ...
export type IndexerSparklineResponseObject = { [marketId: string]: string[] };